requirements.txt
Flask>=3.0
mistralai>=1.12
psycopg2-binary>=2.9
python-dotenv>=1.0